Initial Commit
In the middle of attempting a two container setup. I'm abandoning this for two
reasons.
1) It'll be slow to start a second container for each generate image, meaning
the generate image is more a reference implementation, and a real use would
be a larger script... Less useful, but probably still ok for my use
2) Once you connect to postgres through something other than localhost, you
need to use passwords, I didn't feel like setting that up

Instead, I'll continue using exec to run in the postgres container.